backend/requirements.txt
agno==1.2.15
annotated-types==0.7.0
anyio==4.9.0
beautifulsoup4==4.13.4
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
contourpy==1.3.1
cssselect==1.3.0
cycler==0.12.1
distro==1.9.0
docstring_parser==0.16
duckduckgo_search==8.0.0
fastapi==0.115.12
feedfinder2==0.0.4
feedparser==6.0.11
filelock==3.18.0
fonttools==4.57.0
fsspec==2025.3.2
gitdb==4.0.12
GitPython==3.1.44
google-ai-generativelanguage==0.6.15
google-api-core==2.24.2
google-api-python-client==2.169.0
google-auth==2.40.1
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
google_search_results==2.4.2
googleapis-common-protos==1.70.0
groq==0.22.0
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
huggingface-hub==0.31.1
idna==3.10
jieba3k==0.35.1
Jinja2==3.1.6
joblib==1.5.0
kiwisolver==1.4.8
lxml==5.3.2
lxml_html_clean==0.4.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
mdurl==0.1.2
mpmath==1.3.0
networkx==3.4.2
newspaper3k==0.2.8
nltk==3.9.1
numpy==2.2.4
packaging==24.2
pandas==2.2.3
pillow==11.1.0
primp==0.14.0
proto-plus==1.26.1
protobuf==5.29.4
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.3
pydantic-settings==2.8.1
pydantic_core==2.33.1
Pygments==2.19.1
pyparsing==3.2.3
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-file==2.1.0
rich==14.0.0
rsa==4.9.1
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.3
seaborn==0.13.2
sentence-transformers==4.1.0
setuptools==80.3.1
sgmllib3k==1.0.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
starlette==0.46.1
sympy==1.14.0
threadpoolctl==3.6.0
tinysegmenter==0.3
tldextract==5.3.0
tokenizers==0.21.1
tomli==2.2.1
torch==2.7.0
tqdm==4.67.1
transformers==4.51.3
typer==0.15.2
typing-inspection==0.4.0
typing_extensions==4.13.1
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.4.0
uvicorn==0.34.0
